requirements.txt
Flask
python-dotenv
openai
Flask-Login
Flask-SQLAlchemy 